pyproject.toml
[project]
name = "smart-trader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.9.5",
    "ccxt>=4.3.61",
    "fastapi>=0.111.0",
    "httpx>=0.27.0",
    "matplotlib>=3.9.1",
    "numpy>=2.2.3",
    "pandas>=2.2.2",
    "pydantic>=2.8.2",
    "python-binance>=1.0.28",
    "python-dotenv>=1.0.1",
    "PyYAML>=6.0.1",
    "requests>=2.32.3",
    "seaborn>=0.13.2",
    "ta-lib>=0.6.4",
    "uvicorn>=0.30.1",
    "websocket-client>=1.8.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "flake8>=7.3.0",
    "mypy>=1.17.1",
    "pytest>=8.4.1",
]
